refactor: replace new index store
