+++ spec/store/index.spec.ts
@@ -17,64 +17,166 @@ along with Blendic SVG.  If not, see <https://www.gnu.org/licenses/>.
 Copyright (C) 2021, Tomoya Komiyama.
 */
 
-import { useStore } from '/@/store/index'
-import { getArmature, getBone } from '/@/models'
-import { nextTick } from 'vue'
+import { createStore } from '/@/store/index'
+import { useHistoryStore } from '/@/composables/stores/history'
+import { getBone } from '/@/models'
 
-describe('store/index', () => {
-  const store = useStore()
-  beforeEach(() => {
-    store.state.armatures = []
-    store.state.lastSelectedArmatureId = ''
-    store.state.lastSelectedBoneId = ''
-    store.state.selectedArmatures = {}
-    store.state.selectedBones = {}
-  })
-
-  describe('lastSelectedArmature', () => {
-    it('should clear if the armature is removed', async () => {
-      store.state.armatures = [
-        getArmature({ id: 'a' }),
-        getArmature({ id: 'b' }),
-      ]
-      store.selectArmature('a')
-      expect(store.lastSelectedArmature.value?.id).toBe('a')
-      store.state.armatures = [getArmature({ id: 'a' })]
-      expect(store.lastSelectedArmature.value?.id).toBe('a')
-      store.state.armatures = []
-      expect(store.lastSelectedArmature.value).toBe(undefined)
-      await nextTick()
-      expect(store.state.lastSelectedArmatureId).toBe('')
-    })
-    it('should clear if the armature is unselected', () => {
-      store.state.armatures = [
-        getArmature({ id: 'a' }),
-        getArmature({ id: 'b' }),
-      ]
-      store.selectArmature('a')
-      expect(store.lastSelectedArmature.value?.id).toBe('a')
-      store.state.selectedArmatures = {}
-      expect(store.lastSelectedArmature.value).toBe(undefined)
+describe('src/store/index.ts', () => {
+  describe('addArmature', () => {
+    it('should add new armature', () => {
+      const target = createStore(useHistoryStore())
+      expect(target.armatures.value).toHaveLength(0)
+      target.addArmature('arm_a')
+      expect(target.armatures.value).toHaveLength(1)
+      expect(target.lastSelectedArmatureId.value).toBe('arm_a')
     })
   })
 
-  describe('lastSelectedBone', () => {
-    it('should clear if the bone is removed', async () => {
-      store.state.armatures = [
-        getArmature({
-          id: 'arm',
-          bones: [getBone({ id: 'a' }), getBone({ id: 'b' })],
+  describe('selectAllBone', () => {
+    it('should toggle selected state of all bones', () => {
+      const target = createStore(useHistoryStore())
+      target.addArmature('arm_a')
+      target.addBone('bone_a')
+      target.addBone('bone_b')
+      target.selectAllBone()
+      expect(target.selectedBones.value).toEqual({
+        bone_a: { head: true, tail: true },
+        bone_b: { head: true, tail: true },
+      })
+      target.selectAllBone()
+      expect(target.selectedBones.value).toEqual({})
+    })
+  })
+
+  describe('addBone', () => {
+    it('should add new bone', () => {
+      const target = createStore(useHistoryStore())
+      target.addArmature('arm_a')
+      expect(target.lastSelectedArmature.value!.b_ones).toHaveLength(0)
+      expect(target.boneMap.value).toEqual({})
+      target.addBone('bone_a')
+      target.addBone('bone_b')
+      expect(target.lastSelectedArmature.value!.b_ones).toEqual([
+        'bone_a',
+        'bone_b',
+      ])
+      expect(target.boneMap.value).toEqual({
+        bone_a: expect.anything(),
+        bone_b: expect.anything(),
+      })
+      expect(target.selectedBones.value).toEqual({
+        bone_b: { head: true, tail: true },
+      })
+    })
+  })
+
+  describe('deleteBone', () => {
+    it('should delete bones whose all attrs have been selected and clean connections', () => {
+      const target = createStore(useHistoryStore())
+      target.addArmature('arm_a')
+      target.addBone('bone_a')
+      target.addBone('bone_b')
+      target.selectBones({
+        bone_a: { head: true },
+        bone_b: { head: true, tail: true },
+      })
+      target.deleteBone()
+      expect(target.lastSelectedArmature.value!.b_ones).toEqual(['bone_a'])
+      expect(target.boneMap.value).toEqual({ bone_a: expect.anything() })
+      expect(target.selectedBones.value).toEqual({})
+    })
+  })
+
+  describe('dissolveBone', () => {
+    it('should dissolve bones whose all attrs have been selected and clean connections', () => {
+      const target = createStore(useHistoryStore())
+      target.addArmature('arm_a')
+      target.addBone('bone_a')
+      target.addBone('bone_b')
+      target.addBone('bone_c')
+      target.updateBones({
+        bone_b: { parentId: 'bone_a' },
+        bone_c: { parentId: 'bone_b' },
+      })
+      target.selectBones({
+        bone_a: { head: true },
+        bone_b: { head: true, tail: true },
+      })
+      target.dissolveBone()
+      expect(target.lastSelectedArmature.value!.b_ones).toEqual([
+        'bone_a',
+        'bone_c',
+      ])
+      expect(target.boneMap.value['bone_c'].parentId).toBe('bone_a')
+      expect(target.selectedBones.value).toEqual({})
+    })
+  })
+
+  describe('updateBones', () => {
+    it('should update bones clean connections', () => {
+      const target = createStore(useHistoryStore())
+      target.addArmature('arm_a')
+      target.addBone('bone_a')
+      target.addBone('bone_b')
+      target.addBone('bone_c')
+      target.updateBones({
+        bone_b: { parentId: 'bone_a' },
+        bone_c: { parentId: 'unknown' },
+      })
+      expect(target.boneMap.value['bone_b'].parentId).toBe('bone_a')
+      expect(target.boneMap.value['bone_c'].parentId).toBe('')
+    })
+  })
+
+  describe('updateBone', () => {
+    it('should update bones and fix connections', () => {
+      const target = createStore(useHistoryStore())
+      target.addArmature('arm_a')
+      target.addBone('bone_a')
+      target.addBone('bone_b')
+      target.addBone('bone_c')
+      target.updateBone({
+        parentId: 'unknown',
+      })
+      expect(target.boneMap.value['bone_c'].parentId).toBe('')
+    })
+  })
+
+  describe('updateBoneName', () => {
+    it('should update name of a bone and avoid duplication', () => {
+      const target = createStore(useHistoryStore())
+      target.addArmature('arm_a')
+      target.addBone('bone_a')
+      target.addBone('bone_b')
+      target.selectBone('bone_a')
+      target.updateBoneName('name_a')
+      target.selectBone('bone_b')
+      target.updateBoneName('name_a')
+      expect(target.boneMap.value['bone_a'].name).toBe('name_a')
+      expect(target.boneMap.value['bone_b'].name).toBe('name_a.001')
+    })
+  })
+
+  describe('upsertBones', () => {
+    it('should upsert bones and fix connections', () => {
+      const target = createStore(useHistoryStore())
+      target.addArmature('arm_a')
+      target.addBone('bone_a')
+      target.addBone('bone_b')
+      target.addBone('bone_c')
+      target.upsertBones([
+        getBone({
+          id: 'bone_a',
+          parentId: 'unknown',
+          head: { x: 10, y: 20 },
         }),
-      ]
-      store.selectArmature('arm')
-      store.selectBone('a')
-      expect(store.lastSelectedBone.value?.id).toBe('a')
-      store.state.armatures[0].bones = [getBone({ id: 'a' })]
-      expect(store.lastSelectedBone.value?.id).toBe('a')
-      store.state.armatures[0].bones = []
-      expect(store.lastSelectedBone.value).toBe(undefined)
-      await nextTick()
-      expect(store.state.lastSelectedBoneId).toBe('')
+        getBone({
+          id: 'bone_d',
+        }),
+      ])
+      expect(target.boneMap.value['bone_a'].parentId).toBe('')
+      expect(target.boneMap.value['bone_a'].head).toEqual({ x: 10, y: 20 })
+      expect(target.boneMap.value['bone_d']).toBeDefined()
     })
   })
 })
